Added functions and new elements
I made functions for deleting a job you posted and one for saving jobs to a saved text file, while adding an array in globals.py for that purpose. I also made it so you can view these saved jobs, as well as those jobs someone has already applied to, and added an additional parameter for Jobs in order to specifically delete jobs that you posted, and not others.